refactor: restructure software layers
Refactor router, service, and repository layers:
- Restructure routers, services, and repositories into:
  - layers/{router,service,repository}
- Restructure models and schemas into:
  - schemas/{orm,dto}

Refs: #129
