+++ src/app/schemas/dto/v1/gadgets.py
