+++ src/app/schemas/dto/v1/widgets.py
